use bfloat16 when it is available (namely: not MPS)

+++ nougat/utils/device.py
@@ -2,7 +2,7 @@ import torch
 
 def move_to_device(model):
     if torch.cuda.is_available():
-        return model.to("cuda")
+        return model.to("cuda").to(torch.bfloat16)
     elif torch.backends.mps.is_available():
         return model.to("mps")
-    return model
+    return model.to(torch.bfloat16)
